clean up and seperate code a bit
